yemek-chatbot/requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing
pandas==2.1.1
numpy==1.24.3

# Machine Learning & AI
torch==2.1.0
transformers==4.34.0
sentence-transformers==2.2.2
openai>=1.0.0

# Vector Database
pymilvus==2.3.1

# Memory Management
mem0ai==0.0.8

# Web Framework
fastapi==0.103.2
uvicorn==0.23.2
streamlit==1.27.2

# Utilities
python-dotenv==1.0.0
pydantic==2.4.2
typing-extensions==4.8.0

# Development
pytest==7.4.2
black==23.9.1
flake8==6.1.0

# Configuration
PyYAML==6.0.1